[REFACTOR] Renamed ForgotPasswordForm to ResetPasswordForm and update imports to use ResetPasswordForm in tests and components
